server/requirements.txt
easyocr==1.7.2
Flask==3.1.2
flask_cors==6.0.1
opencv-python-headless==4.12.0.88
python-dotenv==1.2.1